start of css

+++ app/controllers/admins_controller.rb
@@ -3,6 +3,11 @@ class AdminsController < ApplicationController
 
     def index
         @admins = Admin.all
+        if flash[:error]
+            @error = flash[:error]
+        else
+            @error = ""
+        end
     end
 
     def new
@@ -24,9 +29,20 @@ class AdminsController < ApplicationController
     end
 
     def destroy
-        admin = Admin.find(params[:id])
-        admin.destroy
-        redirect_to admins_path
+        if Admin.all.count <=1
+            flash[:error] = "Must have at least 1 admin"
+            redirect_to admins_path
+        else
+            admin = Admin.find(params[:id])
+            if admin.id == sessions[:admin_id]
+                admin.destroy
+                log_out
+                redirect_to '/'
+            else
+                admin.destroy
+                redirect_to admins_path
+            end
+        end
     end
 
     private
